Refactor byte-slab API per review: rename to Points, dedupe layout, move constructors to Points
- Drop PointRef/PointRefIter; Points::iter() yields Result<Point> via the
  existing raw::Point::read_from + Point::new pipeline, collapsing ~170
  lines of duplicated field decoding.
- Move byte-offset layout knowledge into raw::point::Layout as the single
  source of truth, shared between the sequential reader and Points column
  accessors.
- Rename PointCloud -> Points and move reader-backed constructors onto
  the type: Points::from_reader, Points::read_all, Points::fill_from.
  Reader::read_into_cloud / read_all_into_cloud are gone; the internal
  plumbing stays as a pub(crate) helper.
- Benchmark no longer requires the laz feature; default input is
  tests/data/autzen.las, overridable via LAS_BENCH_FILE. Docs point at
  USGS 3DEP / OpenTopography for larger public LAS/LAZ files.
- Fix a pre-existing clippy warning in tests/laszip.rs.
