+++ src/lib.rs
@@ -189,7 +189,7 @@ mod bounds;
 mod color;
 mod error;
 mod gps_time_type;
-mod point_cloud;
+mod points;
 mod transform;
 mod utils;
 mod vector;
@@ -205,7 +205,7 @@ pub use crate::{
     gps_time_type::GpsTimeType,
     header::{Builder, Header},
     point::Point,
-    point_cloud::{PointCloud, PointRef, PointRefIter},
+    points::{PointIter, Points},
     reader::{Reader, ReaderOptions},
     transform::Transform,
     vector::Vector,
